@scott-kausler 2021-05-28
Any idea when this can be merged? The inability to drop a database making cleanup difficult in our CI pipeline.